Rename packages to org.eclipse.microprofile.* domain
