--- proposals/system-test/modules/liberty/build.gradle
@@ -1,5 +1,0 @@
-description = "MicroProfile and JavaEE Testing framework :: Liberty extensions"
-
-dependencies {
-  compile project(':microprofile-system-test-core')
-}
