--- proposals/system-test/modules/liberty/src/main/resources/META-INF/services/org.aguibert.testcontainers.framework.spi.ServerAdapter
@@ -1,1 +1,0 @@
-org.aguibert.testcontainers.framework.liberty.LibertyAdapter
