--- proposals/system-test/sample-apps/everything-app/src/main/java/org/aguibert/liberty/ExternalRestServiceClient.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2019 IBM Corporation and others
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * You may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.aguibert.liberty;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
-
-@ApplicationScoped // needed bean-defining anno for app servers that only support MP Rest Client 1.0
-@RegisterRestClient
-@Path("/mock-passthrough")
-public interface ExternalRestServiceClient {
-
-    @GET
-    @Path("/person/{personId}")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Person getPerson(@PathParam("personId") long id);
-
-}
