+++ proposals/system-test/sample-apps/jdbc-app/src/test/java/org/eclipse/microprofile/system/test/app/DatabaseTest.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright (c) 2019 IBM Corporation and others
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.eclipse.microprofile.system.test.app;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.Collection;
+
+import javax.inject.Inject;
+
+import org.eclipse.microprofile.system.test.app.Person;
+import org.eclipse.microprofile.system.test.app.PersonServiceWithJDBC;
+import org.eclipse.microprofile.system.test.jupiter.MicroProfileTest;
+import org.eclipse.microprofile.system.test.jupiter.SharedContainerConfig;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+@MicroProfileTest
+@SharedContainerConfig(AppContainerConfig.class)
+public class DatabaseTest {
+
+    @Inject
+    public static PersonServiceWithJDBC personSvc;
+
+    @Test
+    public void testGetPerson() {
+        Long bobId = personSvc.createPerson("Bob", 24);
+        Person bob = personSvc.getPerson(bobId);
+        assertEquals("Bob", bob.name);
+        assertEquals(24, bob.age);
+    }
+
+    @Test
+    public void testGetAllPeople() {
+        Long person1Id = personSvc.createPerson("Person1", 1);
+        Long person2Id = personSvc.createPerson("Person2", 2);
+
+        Person expected1 = new Person("Person1", 1, person1Id);
+        Person expected2 = new Person("Person2", 2, person2Id);
+
+        Collection<Person> allPeople = personSvc.getAllPeople();
+        assertTrue(allPeople.size() >= 2, "Expected at least 2 people to be registered, but there were only: " + allPeople);
+        assertTrue(allPeople.contains(expected1), "Did not find person " + expected1 + " in all people: " + allPeople);
+        assertTrue(allPeople.contains(expected2), "Did not find person " + expected2 + " in all people: " + allPeople);
+    }
+
+}
