+++ proposals/system-test/sample-apps/maven-app/src/test/java/org/eclipse/microprofile/system/test/app/it/JaxrsJsonTest.java
@@ -1,0 +1,134 @@
+/*
+ * Copyright (c) 2019 IBM Corporation and others
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.eclipse.microprofile.system.test.app.it;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import java.util.Collection;
+
+import javax.inject.Inject;
+import javax.ws.rs.BadRequestException;
+import javax.ws.rs.NotFoundException;
+
+import org.aguibert.testcontainers.framework.MicroProfileApplication;
+import org.eclipse.microprofile.system.test.app.Person;
+import org.eclipse.microprofile.system.test.app.PersonService;
+import org.eclipse.microprofile.system.test.jupiter.MicroProfileTest;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+@MicroProfileTest
+public class JaxrsJsonTest {
+    
+    @Container
+    public static MicroProfileApplication<?> app = new MicroProfileApplication<>()
+                    .withAppContextRoot("/myservice")
+                    .withReadinessPath("/myservice/people");
+
+    @Inject
+    public static PersonService personSvc;
+
+    @Test
+    public void testCreatePerson() {
+        Long createId = personSvc.createPerson("Hank", 42);
+        assertNotNull(createId);
+    }
+
+    @Test
+    public void testMinSizeName() {
+        Long minSizeNameId = personSvc.createPerson("Ha", 42);
+        assertEquals(new Person("Ha", 42, minSizeNameId),
+                     personSvc.getPerson(minSizeNameId));
+    }
+
+    @Test
+    public void testMinAge() {
+        Long minAgeId = personSvc.createPerson("Newborn", 0);
+        assertEquals(new Person("Newborn", 0, minAgeId),
+                     personSvc.getPerson(minAgeId));
+    }
+
+    @Test
+    public void testGetPerson() {
+        Long bobId = personSvc.createPerson("Bob", 24);
+        Person bob = personSvc.getPerson(bobId);
+        assertEquals("Bob", bob.name);
+        assertEquals(24, bob.age);
+        assertNotNull(bob.id);
+    }
+
+    @Test
+    public void testGetAllPeople() {
+        Long person1Id = personSvc.createPerson("Person1", 1);
+        Long person2Id = personSvc.createPerson("Person2", 2);
+
+        Person expected1 = new Person("Person1", 1, person1Id);
+        Person expected2 = new Person("Person2", 2, person2Id);
+
+        Collection<Person> allPeople = personSvc.getAllPeople();
+        assertTrue("Expected at least 2 people to be registered, but there were only: " + allPeople,
+                   allPeople.size() >= 2);
+        assertTrue("Did not find person " + expected1 + " in all people: " + allPeople,
+                   allPeople.contains(expected1));
+        assertTrue("Did not find person " + expected2 + " in all people: " + allPeople,
+                   allPeople.contains(expected2));
+    }
+
+    @Test
+    public void testUpdateAge() {
+        Long personId = personSvc.createPerson("newAgePerson", 1);
+
+        Person originalPerson = personSvc.getPerson(personId);
+        assertEquals("newAgePerson", originalPerson.name);
+        assertEquals(1, originalPerson.age);
+        assertEquals(personId, Long.valueOf(originalPerson.id));
+
+        personSvc.updatePerson(personId, new Person(originalPerson.name, 2, originalPerson.id));
+        Person updatedPerson = personSvc.getPerson(personId);
+        assertEquals("newAgePerson", updatedPerson.name);
+        assertEquals(2, updatedPerson.age);
+        assertEquals(personId, Long.valueOf(updatedPerson.id));
+    }
+
+    @Test
+    public void testGetUnknownPerson() {
+        assertThrows(NotFoundException.class, () -> personSvc.getPerson(-1L));
+    }
+
+    @Test
+    public void testCreateBadPersonNullName() {
+        assertThrows(BadRequestException.class, () -> personSvc.createPerson(null, 5));
+    }
+
+    @Test
+    public void testCreateBadPersonNegativeAge() {
+        assertThrows(BadRequestException.class, () -> personSvc.createPerson("NegativeAgePersoN", -1));
+    }
+
+    @Test
+    public void testCreateBadPersonNameTooLong() {
+        assertThrows(BadRequestException.class, () -> personSvc.createPerson("NameTooLongPersonNameTooLongPersonNameTooLongPerson", 5));
+    }
+
+}
